Extract match test to own file and add simplified versions
