+++ cpg-language-python/src/test/kotlin/de/fraunhofer/aisec/cpg/frontends/python/statementHandler/StatementHandlerTest.kt
@@ -60,102 +60,6 @@ class StatementHandlerTest : BaseTest() {
     }
 
     @Test
-    fun testMatch2() {
-        analyzeFile("match.py")
-
-        val func = result.functions["match_weird"]
-        assertNotNull(func)
-
-        val switchStatement = func.switches.singleOrNull()
-        assertNotNull(switchStatement)
-
-        assertIs<CallExpression>(switchStatement.selector)
-
-        val statementBlock = switchStatement.statement
-        assertIs<Block>(statementBlock)
-        val case = statementBlock[0]
-        assertIs<CaseStatement>(case)
-        assertIs<ProblemExpression>(case.caseExpression)
-    }
-
-    @Test
-    fun testMatch() {
-        analyzeFile("match.py")
-
-        val func = result.functions["matcher"]
-        assertNotNull(func)
-
-        val switchStatement = func.switches.singleOrNull()
-        assertNotNull(switchStatement)
-
-        assertLocalName("x", switchStatement.selector)
-        assertIs<Reference>(switchStatement.selector)
-        val paramX = func.parameters.singleOrNull()
-        assertNotNull(paramX)
-        assertRefersTo(switchStatement.selector, paramX)
-
-        val statementBlock = switchStatement.statement
-        assertIs<Block>(statementBlock)
-        val caseSingleton = statementBlock[0]
-        assertIs<CaseStatement>(caseSingleton)
-        val singletonCheck = caseSingleton.caseExpression
-        assertIs<BinaryOperator>(singletonCheck)
-        assertEquals("===", singletonCheck.operatorCode)
-        assertRefersTo(singletonCheck.lhs, paramX)
-        val singletonRhs = singletonCheck.rhs
-        assertIs<Literal<*>>(singletonRhs)
-        assertNull(singletonRhs.value)
-        assertIs<BreakStatement>(statementBlock[2])
-
-        val caseValue = statementBlock[3]
-        assertIs<CaseStatement>(caseValue)
-        val valueCheck = caseValue.caseExpression
-        assertIs<BinaryOperator>(valueCheck)
-        assertEquals("==", valueCheck.operatorCode)
-        assertRefersTo(valueCheck.lhs, paramX)
-        assertLiteralValue("value", valueCheck.rhs)
-        assertIs<BreakStatement>(statementBlock[5])
-
-        val caseAnd = statementBlock[6]
-        assertIs<CaseStatement>(caseAnd)
-        val andExpr = caseAnd.caseExpression
-        assertIs<BinaryOperator>(andExpr)
-        assertEquals("and", andExpr.operatorCode)
-        val andRhs = andExpr.rhs
-        assertIs<BinaryOperator>(andRhs)
-        assertEquals(">", andRhs.operatorCode)
-        assertRefersTo(andRhs.lhs, paramX)
-        assertLiteralValue(0L, andRhs.rhs)
-        assertIs<BreakStatement>(statementBlock[8])
-
-        assertIs<CaseStatement>(statementBlock[9])
-        assertIs<BreakStatement>(statementBlock[11])
-        assertIs<CaseStatement>(statementBlock[12])
-        assertIs<BreakStatement>(statementBlock[14])
-        assertIs<CaseStatement>(statementBlock[15])
-        assertIs<BreakStatement>(statementBlock[17])
-        assertIs<CaseStatement>(statementBlock[18])
-        assertIs<BreakStatement>(statementBlock[20])
-        assertIs<CaseStatement>(statementBlock[21])
-        assertIs<BreakStatement>(statementBlock[23])
-        assertIs<CaseStatement>(statementBlock[24])
-        assertIs<BreakStatement>(statementBlock[26])
-
-        val caseOr = statementBlock[27]
-        assertIs<CaseStatement>(caseOr)
-        val orExpr = caseOr.caseExpression
-        assertIs<BinaryOperator>(orExpr)
-        assertEquals("or", orExpr.operatorCode)
-        assertIs<BinaryOperator>(orExpr.lhs)
-        assertIs<BinaryOperator>(orExpr.rhs)
-        assertIs<BreakStatement>(statementBlock[29])
-
-        val caseDefault = statementBlock[30]
-        assertIs<DefaultStatement>(caseDefault)
-        assertIs<BreakStatement>(statementBlock[32])
-    }
-
-    @Test
     fun testTry() {
         val tu =
             analyzeAndGetFirstTU(listOf(topLevel.resolve("try.py").toFile()), topLevel, true) {
